refactor: make main.go smaller
